task6-refactor/requirements.txt
# requirements.txt
torch>=1.10.0
transformers>=4.30.0
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
pyyaml>=5.4.0
optuna>=2.10.0
nltk>=3.7